Move helpers to util module
