+++ lua/quicktest/adapters/meson/init.lua
@@ -1,5 +1,5 @@
 local Job = require("plenary.job")
-local json5 = require("json5")
+local util = require("quicktest.adapters.meson.util")
 
 local M = {
   name = "meson test runner for C assuming Criterion test frame work",
@@ -7,117 +7,8 @@ local M = {
 
 local parsed_test_output = {}
 
-local function capture_json(data, json)
-  local complete = false
-
-  if vim.startswith(data, "{") then
-    json.open = true
-    json.text = ""
-  end
-
-  if json.open then
-    json.text = json.text .. data .. "\n"
-  end
-
-  if vim.startswith(data, "}") then
-    json.open = false
-    complete = true
-  end
-
-  return complete
-end
-
-local function print_results(data, params, send)
-  local parsed = json5.parse(data)
-  -- print(vim.inspect(parsed))
-  for _, ts in ipairs(parsed["test_suites"]) do
-    -- print(vim.inspect(ts))
-    if params.test.test_suite == nil or params.test.test_suite == ts["name"] then
-      for _, test in ipairs(ts["tests"]) do
-        -- print(vim.inspect(test))
-        if params.test.test_name == nil or params.test.test_name == test["name"] then
-          send({ type = "stdout", output = test["name"] .. ": " .. test["status"] })
-          if test["messages"] then
-            for _, msg in ipairs(test["messages"]) do
-              send({ type = "stdout", output = "  " .. msg })
-            end
-          end
-        end
-      end
-    end
-  end
-end
-
-local function capture_xml(data, xml)
-  local returned_data = data
-
-  if vim.startswith(data, "<?xml") then
-    xml.open = true
-    xml.text = ""
-  end
-
-  if xml.open then
-    xml.text = xml.text .. data .. "\n"
-    returned_data = nil
-  end
-
-  if vim.startswith(data, "</testsuites>") then
-    xml.open = false
-  end
-
-  return returned_data
-end
-
-local function get_filename(path)
-  return path:match("[^/]*.c$")
-end
-
-local function mysplit(inputstr, sep)
-  if sep == nil then
-    sep = "%s"
-  end
-  local t = {}
-  for str in string.gmatch(inputstr, "([^" .. sep .. "]+)") do
-    table.insert(t, str)
-  end
-  return t
-end
-
--- Assumes the test executable is named the same as the test except for the file ending
--- e.g. test_name.c -> test_name
-local function get_test_exe_from_buffer(bufnr)
-  local bufname = vim.api.nvim_buf_get_name(bufnr)
-  local filename = get_filename(bufname)
-  filename = string.gsub(filename, ".c", "", 1)
-  return filename
-end
-
-local function get_test_suite_and_name(line)
-  local parts = mysplit(line, "(")
-  local index = 1
-
-  if vim.startswith(parts[1], "ParameterizedTest") then
-    index = 2
-  end
-
-  parts = mysplit(parts[2], ",")
-  return {
-    test_suite = string.gsub(parts[index], "%s+", ""),
-    test_name = string.gsub(parts[index + 1], "%s+", ""),
-  }
-end
-
-local function get_nearest_test(bufnr, cursor_pos)
-  for pos = cursor_pos[1], 1, -1 do
-    local line = vim.api.nvim_buf_get_lines(bufnr, pos - 1, pos, true)[1]
-    if vim.startswith(line, "Test(") or vim.startswith(line, "ParameterizedTest(") then
-      return line
-    end
-  end
-end
-
 M.build_file_run_params = function(bufnr, cursor_pos)
-  local test_exe = get_test_exe_from_buffer(bufnr)
+  local test_exe = util.get_test_exe_from_buffer(bufnr)
   return {
     test = {},
     test_exe = test_exe,
@@ -127,9 +18,9 @@ M.build_file_run_params = function(bufnr, cursor_pos)
 end
 
 M.build_line_run_params = function(bufnr, cursor_pos)
-  local test_exe = get_test_exe_from_buffer(bufnr)
-  local line = get_nearest_test(bufnr, cursor_pos)
-  local test = get_test_suite_and_name(line)
+  local test_exe = util.get_test_exe_from_buffer(bufnr)
+  local line = util.get_nearest_test(bufnr, cursor_pos)
+  local test = util.get_test_suite_and_name(line)
   return {
     test = test,
     test_exe = test_exe,
@@ -150,56 +41,17 @@ end
 ---@param send fun(data: any)
 ---@return integer
 M.run = function(params, send)
-  local test_args = { "test", "-C", "build" }
-
-  if params.test_exe then
-    table.insert(test_args, params.test_exe)
-  end
-
-  local ts = params.test.test_suite or "*"
-  local tn = params.test.test_name or "*"
-  table.insert(test_args, "--test-args=--filter=" .. ts .. "/" .. tn)
-
-  table.insert(test_args, "--test-args=--json")
-  table.insert(test_args, "-v")
-
-  print(vim.inspect(test_args))
-
-  local build_ok = false
-  local build_out = {}
-  local build = Job:new({
-    command = "meson",
-    args = { "compile", "-C", "build" },
-    on_stdout = function(_, data)
-      table.insert(build_out, data)
-    end,
-    on_stderr = function(_, data)
-      send({ type = "stderr", output = data })
-    end,
-    on_exit = function(_, return_val)
-      build_ok = return_val == 0
-      if build_ok == false then
-        for _, value in ipairs(build_out) do
-          send({ type = "stdout", output = value })
-        end
-        send({ type = "exit", code = return_val })
-      end
-    end,
-  })
-  build:start()
-  Job.join(build)
-
-  if build_ok == false then
+  if util.build_tests(send) == false then
     return -1
   end
 
   local job = Job:new({
     command = "meson",
-    args = test_args, -- Modify based on how your test command needs to be structured
+    args = util.make_test_args(params),
     on_stdout = function(_, data)
-      data = capture_json(data, parsed_test_output)
+      data = util.capture_json(data, parsed_test_output)
       if data then
-        print_results(parsed_test_output.text, params, send)
+        util.print_results(parsed_test_output.text, params, send)
       end
     end,
     on_stderr = function(_, data)
@@ -224,7 +76,6 @@ M.after_run = function(params, results)
     return
   end
   -- print(xml_output.text)
-  local parsed_data = json5.parse(parsed_test_output.text)
   -- print(vim.inspect(parsed_data))
   -- print(parsed_data["id"])
   -- print(parsed_data["test_suites"][1]["name"])
@@ -235,7 +86,7 @@ end
 ---@return boolean
 M.is_enabled = function(bufnr)
   local bufname = vim.api.nvim_buf_get_name(bufnr)
-  local filename = get_filename(bufname)
+  local filename = util.get_filename(bufname)
   return vim.startswith(filename, "test_") and vim.endswith(filename, ".c")
 end
